// Samples/GoodToDo/GoodToDo.ClientApp/src/apiclient/ApiClient.ts
export class Result implements IResult {
    success!: boolean;
    errors?: ErrorDto[] | undefined;
    failure!: boolean;

    constructor(data?: IResult) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.success = _data["success"];
            if (Array.isArray(_data["errors"])) {
                this.errors = [] as any;
                for (let i in _data["errors"])
                    this.errors!.push(ErrorDto.fromJS(_data["errors"][i]));
            }
            this.failure = _data["failure"];
        }
    }

    static fromJS(data: any): Result {
        data = typeof data === 'object' ? data : {};
        let result = new Result();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["success"] = this.success;
        if (Array.isArray(this.errors)) {
            data["errors"] = [];
            for (let i in this.errors)
                data["errors"].push(this.errors[i].toJSON());
        }
        data["failure"] = this.failure;
        return data; 
    }
}

export interface IResult {
    success: boolean;
    errors?: ErrorDto[] | undefined;
    failure: boolean;
}

export class ErrorDto implements IErrorDto {
    key?: string | undefined;
    message?: string | undefined;

    constructor(data?: IErrorDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.key = _data["key"];
            this.message = _data["message"];
        }
    }

    static fromJS(data: any): ErrorDto {
        data = typeof data === 'object' ? data : {};
        let result = new ErrorDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["key"] = this.key;
        data["message"] = this.message;
        return data; 
    }
}

export interface IErrorDto {
    key?: string | undefined;
    message?: string | undefined;
}
// ...
export class EntityOfGuid implements IEntityOfGuid {
    id!: string;

    constructor(data?: IEntityOfGuid) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
        }
    }

    static fromJS(data: any): EntityOfGuid {
        data = typeof data === 'object' ? data : {};
        let result = new EntityOfGuid();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        return data; 
    }
}

export interface IEntityOfGuid {
    id: string | undefined;
}

export class ToDoItem extends EntityOfGuid implements IToDoItem {
    userId!: string;
    message!: string;
    complete!: boolean;

    constructor(data?: IToDoItem) {
        super(data);
    }

    init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.userId = _data["userId"];
            this.message = _data["message"];
            this.complete = _data["complete"];
        }
    }

    static fromJS(data: any): ToDoItem {
        data = typeof data === 'object' ? data : {};
        let result = new ToDoItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["userId"] = this.userId;
        data["message"] = this.message;
        data["complete"] = this.complete;
        super.toJSON(data);
        return data; 
    }
}

export interface IToDoItem extends IEntityOfGuid {
    userId: string;
    message: string;
    complete: boolean;
}

export class ApiException extends Error {
    message: string;
    status: number; 
    response: string; 
    headers: { [key: string]: any; };
    result: any; 

    constructor(message: string, status: number, response: string, headers: { [key: string]: any; }, result: any) {
        super();

        this.message = message;
        this.status = status;
        this.response = response;
        this.headers = headers;
        this.result = result;
    }

    protected isApiException = true;

    static isApiException(obj: any): obj is ApiException {
        return obj.isApiException === true;
    }
}